[#122] feat: 수정 - auth middleware를 거치도록 api 수정

+++ server/src/controllers/cartController.ts
@@ -10,7 +10,7 @@ const create = async (
   res: Response,
   next: NextFunction
 ): Promise<void> => {
-  const cart = await Cart.create(req.body);
+  const cart = await Cart.create({ ...req.body, userId: req.user as number });
   res.status(200).send(cart);
 };
 
@@ -21,7 +21,8 @@ const remove = async (
 ): Promise<void> => {
   const { params } = req;
   const productId = parseInt(params.productId, 10);
-  await Cart.destroy({ where: { productId } });
+  const userId = req.user as number;
+  await Cart.destroy({ where: { productId, userId } });
   res.status(200).send({ success: true });
 };
 
@@ -30,9 +31,9 @@ const find = async (
   res: Response,
   next: NextFunction
 ): Promise<void> => {
-  const { params } = req;
-  const userId = parseInt(params.userId, 10);
-  const result = await CartService.findCartIncludeProductByUserId(userId);
+  const result = await CartService.findCartIncludeProductByUserId(
+    req.user as number
+  );
   const carts = result.map((row) => {
     // eslint-disable-next-line no-shadow
     const { userId, count, ...product } = row;
@@ -46,7 +47,8 @@ const update = async (
   res: Response,
   next: NextFunction
 ): Promise<void> => {
-  const { userId, productId, count } = req.body;
+  const { productId, count } = req.body;
+  const userId = req.user as number;
   await Cart.update({ count }, { where: { userId, productId } });
   res.status(200).send({ success: true });
 };
